--- app/api/__init__.py
@@ -1,8 +1,0 @@
-from fastapi import FastAPI, APIRouter
-
-app = FastAPI(root_path="/api")
-
-
-@app.get("/")
-async def root():
-    return {"message": "Hello World"}
