--- app/api/dependency/__init__.py
@@ -1,1 +1,0 @@
-from .user import get_current_user, get_user
